Tmp build fixes.

+++ src/util/math_cdf.cpp
@@ -23,14 +23,14 @@ void util_cdf_invert(const int resolution,
     const int half_size = (resolution - 1) / 2;
     for (int i = 0; i <= half_size; i++) {
       float x = i / (float)half_size;
-      int index = upper_bound(cdf.begin(), cdf.end(), x) - cdf.begin();
+      int index = (int)(upper_bound(cdf.begin(), cdf.end(), x) - cdf.begin());
       float t;
       if (index < cdf.size() - 1) {
         t = (x - cdf[index]) / (cdf[index + 1] - cdf[index]);
       }
       else {
         t = 0.0f;
-        index = cdf.size() - 1;
+        index = (int)cdf.size() - 1;
       }
       float y = ((index + t) / (resolution - 1)) * (2.0f * range);
       inv_cdf[half_size + i] = 0.5f * (1.0f + y);
@@ -40,7 +40,7 @@ void util_cdf_invert(const int resolution,
   else {
     for (int i = 0; i < resolution; i++) {
       float x = from + range * (float)i * inv_resolution;
-      int index = upper_bound(cdf.begin(), cdf.end(), x) - cdf.begin();
+      int index = (int)(upper_bound(cdf.begin(), cdf.end(), x) - cdf.begin());
       float t;
       if (index < cdf.size() - 1) {
         t = (x - cdf[index]) / (cdf[index + 1] - cdf[index]);
